Simplify `Distribute` by retaining only required services
Split module into several files
